+++ packages/engine/Source/DataSources/CompositeProperty.js
@@ -1,5 +1,5 @@
+import { JulianDate } from "@cesium/engine";
 import defined from "../Core/defined.js";
-import DeveloperError from "../Core/DeveloperError.js";
 import Event from "../Core/Event.js";
 import EventHelper from "../Core/EventHelper.js";
 import TimeIntervalCollection from "../Core/TimeIntervalCollection.js";
@@ -106,16 +106,14 @@ Object.defineProperties(CompositeProperty.prototype, {
 /**
  * Gets the value of the property at the provided time.
  *
- * @param {JulianDate} time The time for which to retrieve the value.
+ * @param {JulianDate} [time=JulianDate.now()] The time for which to retrieve the value. If omitted, the current system time is used.
  * @param {object} [result] The object to store the value into, if omitted, a new instance is created and returned.
  * @returns {object} The modified result parameter or a new instance if the result parameter was not supplied.
  */
 CompositeProperty.prototype.getValue = function (time, result) {
-  //>>includeStart('debug', pragmas.debug);
   if (!defined(time)) {
-    throw new DeveloperError("time is required");
+    time = JulianDate.now();
   }
-  //>>includeEnd('debug');
 
   const innerProperty = this._intervals.findDataForIntervalContainingDate(time);
   if (defined(innerProperty)) {